requirements.txt
requests
delegator.py